War of the WaRnINGs Part the Ninth, Chapter F: All tests passed
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta/@54622 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: a22304bc2f804ca402517ef531e2fd3da901dc32

+++ rosetta_source/src/protocols/forge/remodel/RemodelData.cc
@@ -290,7 +290,7 @@ void RemodelData::getLoopsToBuildFromFile( std::string filename ) {
 								else {
 													oss << line.index + length*rep << " _ " ;
 								}
-					for ( Size i = 3; i< (int)split_info.size(); i++ ) {
+					for ( Size i = 3; i< split_info.size(); i++ ) {
 						if (split_info[i].substr(0,3) != "CST"){
 							oss << split_info[i] << " " ;
 						}
